refactoring all main related fragments into activity package for better oversight and clarity
