+++ app/src/main/java/com/example/fithub/main/AnalysisFragment.java
@@ -1,4 +1,4 @@
-package com.example.fithub;
+package com.example.fithub.main;
 
 import android.graphics.Color;
 import android.os.Bundle;
@@ -10,6 +10,7 @@ import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.navigation.fragment.NavHostFragment;
 
+import com.example.fithub.R;
 import com.example.fithub.databinding.FragmentAnalysisBinding;
 import com.github.mikephil.charting.charts.PieChart;
 import com.github.mikephil.charting.data.PieData;
