+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,17 +7,12 @@
 
     <fragment
         android:id="@+id/FirstFragment"
-        android:name="com.example.fithub.FirstFragment"
+        android:name="com.example.fithub.main.FirstFragment"
         android:label="@string/home_fragment_label"
-        tools:layout="@layout/fragment_first">
-
-        <action
-            android:id="@+id/action_FirstFragment_to_SecondFragment"
-            app:destination="@id/AnalysisFragment" />
-    </fragment>
+        tools:layout="@layout/fragment_first"/>
     <fragment
         android:id="@+id/SecondFragment"
-        android:name="com.example.fithub.SecondFragment"
+        android:name="com.example.fithub.main.SecondFragment"
         android:label="@string/second_fragment_label"
         tools:layout="@layout/fragment_second">
 
@@ -25,9 +20,4 @@
             android:id="@+id/action_SecondFragment_to_FirstFragment"
             app:destination="@id/FirstFragment" />
     </fragment>
-    <fragment
-        android:id="@+id/AnalysisFragment"
-        android:name="com.example.fithub.AnalysisFragment"
-        android:label="@string/analysis_fragment_label"
-        tools:layout="@layout/fragment_analysis"/>
 </navigation>
